+++ src/components/Header.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import { Link } from 'react-router-dom';
+
+const header = () => {
+  return (
+    <div>
+      <Link to='/'>Home</Link>
+      <br/>
+      <Link to='/login'>Login</Link>
+    </div>
+  );
+};
+
+export default header;
