+++ src/pages/Login/index.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import Header from '../../components/Header';
+
+const login = () => {
+  return (
+    <div>
+      <Header />
+      <div>Login</div>
+      <button onClick={() => {alert('登录')}}>登录</button>
+    </div>
+  );
+};
+
+export default login;
